requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
requests>=2.26.0
beautifulsoup4>=4.9.3
selenium>=4.0.0
webdriver_manager>=3.8.0
python-multipart>=0.0.5
typing-extensions>=4.0.0
aiofiles>=0.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sentence-transformers>=2.2.2
torch>=2.0.0
easyocr>=1.7.0
numpy>=1.21.0
Pillow>=8.3.2
scikit-learn>=1.0.2